feat: 제네릭을 이용한 실수(double) 입력 구현 및 연산 타입 double로 변경, 숫자 포맷 출력 기능 추가

+++ src/main/java/lv3/OperatorType.java
@@ -16,13 +16,20 @@ public enum OperatorType {
 
 
     private final char symbol; // 수정: 쉼표를 세미콜론으로 변경
-    private final BiFunction<Integer, Integer, Integer> operation;
+    private final BiFunction<Double, Double, Double> operation;
 
-    OperatorType(char command, BiFunction<Integer, Integer, Integer> operation) {
+    OperatorType(char command, BiFunction<Double, Double, Double> operation) {
         this.symbol = command;
         this.operation = operation;
     }
 
+
+    /**
+     * 사용자로부터 입력받은 기호에 해당하는 operator를 반환합니다.
+     *
+     * @param command
+     * @return
+     */
     public static OperatorType getOperatorFromCmd(char command) {
         return Arrays.stream(values())
                 .filter(op -> op.symbol == command)
@@ -30,7 +37,15 @@ public enum OperatorType {
                 .orElseThrow(() -> new IllegalArgumentException("오류: 지원하지 않는 연산자 기호입니다: " + command));
     }
 
-    public int operate(int first, int second) {
+
+    /**
+     * operator 유형별로 BiFunction 함수형 인터페이스의 메소드인 apply에 람다로 정의한 연산을 수행하도록 합니다.
+     *
+     * @param first
+     * @param second
+     * @return
+     */
+    public double operate(double first, double second) {
         return operation.apply(first, second);
     }
